Add files via upload
Before uploading this Power App into your environment you will need to ensure the user uploading the package via the "Import canvas app" link in the Power Apps web studio has a Power Apps Plan 2 license or Trial license assigned to that user. 

If not, when importing the app you will likely receive an error message indicating the 2 flows associated to the app were not imported, with a rather obsure error message which (if you can find) will inform you that user does not have a license plan necessary to import the app. 

Per a thread on the Power Apps community site you can obtain a 90 day trial license in order to test and evaluate this app in your Power Apps environment per the following links for additional insight (and to sign up for the trial license):
https://powerusers.microsoft.com/t5/Building-Power-Apps/What-is-quot-Microsoft-PowerApps-Trial-Plan-2-quot/m-p/131622#M45812
https://powerapps.microsoft.com/en-us/#